Remove theme validation, Refactor and reorder plugin validation script.
