--- .github/workflows/validate-theme-entry.yml
@@ -1,28 +1,0 @@
-name: Validate Theme Entry
-
-on:
-  pull_request_target:
-    branches:
-      - master
-    paths:
-      - community-css-themes.json
-
-jobs:
-  theme-validation:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          ref: "refs/pull/${{ github.event.number }}/merge"
-      - uses: actions/setup-node@v2
-      - run: npm install probe-image-size
-      - uses: actions/github-script@v6
-        with:
-          script: |
-            const script = require('./validation-theme.js');
-            const probe = require('probe-image-size');
-            return await script({github, context, core, probe});
-    permissions:
-      contents: write
-      issues: write
-      pull-requests: write
